--- a/hypothesis-python/src/hypothesis/internal/conjecture/providers.py
+++ b/hypothesis-python/src/hypothesis/internal/conjecture/providers.py
@@ -12,6 +12,7 @@ import abc
 import contextlib
 import math
 from collections.abc import Iterable
+from random import Random
 from sys import float_info
 from typing import (
     TYPE_CHECKING,
@@ -348,20 +349,20 @@ class HypothesisProvider(PrimitiveProvider):
 
     def __init__(self, conjecturedata: Optional["ConjectureData"], /):
         super().__init__(conjecturedata)
+        self._random = None if self._cd is None else self._cd._random
 
     def draw_boolean(
         self,
         p: float = 0.5,
     ) -> bool:
-        assert self._cd is not None
-        assert self._cd._random is not None
+        assert self._random is not None
 
         if p <= 0:
             return False
         if p >= 1:
             return True
 
-        return self._cd._random.random() < p
+        return self._random.random() < p
 
     def draw_integer(
         self,
@@ -471,7 +472,7 @@ class HypothesisProvider(PrimitiveProvider):
         max_size: int = COLLECTION_DEFAULT_MAX_SIZE,
     ) -> str:
         assert self._cd is not None
-        assert self._cd._random is not None
+        assert self._random is not None
 
         if len(intervals) == 0:
             return ""
@@ -501,11 +502,11 @@ class HypothesisProvider(PrimitiveProvider):
         while elements.more():
             if len(intervals) > 256:
                 if self.draw_boolean(0.2):
-                    i = self._cd._random.randint(256, len(intervals) - 1)
+                    i = self._random.randint(256, len(intervals) - 1)
                 else:
-                    i = self._cd._random.randint(0, 255)
+                    i = self._random.randint(0, 255)
             else:
-                i = self._cd._random.randint(0, len(intervals) - 1)
+                i = self._random.randint(0, len(intervals) - 1)
 
             chars.append(intervals.char_in_shrink_order(i))
 
@@ -517,7 +518,7 @@ class HypothesisProvider(PrimitiveProvider):
         max_size: int = COLLECTION_DEFAULT_MAX_SIZE,
     ) -> bytes:
         assert self._cd is not None
-        assert self._cd._random is not None
+        assert self._random is not None
 
         buf = bytearray()
         average_size = min(
@@ -532,25 +533,24 @@ class HypothesisProvider(PrimitiveProvider):
             observe=False,
         )
         while elements.more():
-            buf += self._cd._random.randbytes(1)
+            buf += self._random.randbytes(1)
 
         return bytes(buf)
 
     def _draw_float(self) -> float:
-        assert self._cd is not None
-        assert self._cd._random is not None
+        assert self._random is not None
 
-        f = lex_to_float(self._cd._random.getrandbits(64))
-        sign = 1 if self._cd._random.getrandbits(1) else -1
+        f = lex_to_float(self._random.getrandbits(64))
+        sign = 1 if self._random.getrandbits(1) else -1
         return sign * f
 
     def _draw_unbounded_integer(self) -> int:
         assert self._cd is not None
-        assert self._cd._random is not None
+        assert self._random is not None
 
         size = INT_SIZES[INT_SIZES_SAMPLER.sample(self._cd)]
 
-        r = self._cd._random.getrandbits(size)
+        r = self._random.getrandbits(size)
         sign = r & 1
         r >>= 1
         if sign:
@@ -566,22 +566,22 @@ class HypothesisProvider(PrimitiveProvider):
     ) -> int:
         assert lower <= upper
         assert self._cd is not None
-        assert self._cd._random is not None
+        assert self._random is not None
 
         if lower == upper:
             return lower
 
         bits = (upper - lower).bit_length()
-        if bits > 24 and vary_size and self._cd._random.random() < 7 / 8:
+        if bits > 24 and vary_size and self._random.random() < 7 / 8:
             # For large ranges, we combine the uniform random distribution
             # with a weighting scheme with moderate chance.  Cutoff at 2 ** 24 so that our
             # choice of unicode characters is uniform but the 32bit distribution is not.
             idx = INT_SIZES_SAMPLER.sample(self._cd)
             cap_bits = min(bits, INT_SIZES[idx])
             upper = min(upper, lower + 2**cap_bits - 1)
-            return self._cd._random.randint(lower, upper)
+            return self._random.randint(lower, upper)
 
-        return self._cd._random.randint(lower, upper)
+        return self._random.randint(lower, upper)
 
     @classmethod
     def _draw_float_init_logic(
@@ -827,3 +827,30 @@ class BytestringProvider(PrimitiveProvider):
     ) -> bytes:
         values = self._draw_collection(min_size, max_size, alphabet_size=2**8)
         return bytes(values)
+
+
+class URandom(Random):
+    # we reimplement a Random instance instead of using SystemRandom, because
+    # os.urandom is not guaranteed to read from /dev/urandom.
+    def getrandbits(self, k: int) -> int:
+        assert k >= 0
+        size = bits_to_bytes(k)
+        with open("/dev/urandom", "rb") as f:
+            n = int_from_bytes(bytearray(f.read(size)))
+        # trim excess bits
+        return n >> (size * 8 - k)
+
+
+class URandomProvider(HypothesisProvider):
+    # A provider which reads directly from /dev/urandom as its source of randomness.
+    # This provider exists to provide better Hypothesis integration with Antithesis
+    # (https://antithesis.com/), which interprets calls to /dev/urandom as the
+    # randomness to mutate. This effectively gives Antithesis control over
+    # the choices made by the URandomProvider.
+    #
+    # If you are not using Antithesis, you probably don't want to use this
+    # provider.
+
+    def __init__(self, conjecturedata: Optional["ConjectureData"], /):
+        super().__init__(conjecturedata)
+        self._random = URandom()