--- a/hypothesis-python/src/hypothesis/internal/conjecture/providers.py
+++ b/hypothesis-python/src/hypothesis/internal/conjecture/providers.py
@@ -868,7 +868,7 @@ class URandomProvider(HypothesisProvider):
 
     def __init__(self, conjecturedata: Optional["ConjectureData"], /):
         super().__init__(conjecturedata)
-        if WINDOWS:
+        if WINDOWS:  # pragma: no branch
             warnings.warn(
                 "/dev/urandom is not available on windows. Falling back to "
                 'standard PRNG generation (equivalent to backend="hypothesis").',