--- a/hypothesis-python/tests/conjecture/test_alt_backend.py
+++ b/hypothesis-python/tests/conjecture/test_alt_backend.py
@@ -603,6 +603,14 @@ def test_available_providers_deprecation():
     "strategy", [st.integers(), st.text(), st.floats(), st.booleans(), st.binary()]
 )
 def test_can_generate_from_all_available_providers(backend, strategy):
+    if backend == "crosshair":
+        # TODO running into a 'not in statespace' issue which is fixed in
+        # https://github.com/HypothesisWorks/hypothesis/pull/4034. Remove
+        # this skip when that is merged"
+        pytest.skip(
+            "temp, fixed in https://github.com/HypothesisWorks/hypothesis/pull/4034"
+        )
+
     if backend == "hypothesis-urandom" and WINDOWS:
         pytest.skip("/dev/urandom not available on windows")