[Antithesis](https://antithesis.com/) is a fuzzer which uses `/dev/urandom` as its source of randomness to mutate. This pull adds a `class URandomProvider(HypothesisProvider)` which backs its random draws using `/dev/urandom`, giving Antithesis control over choices made during Hypothesis tests.

(Zac and I are attending [Antithesis' BugBash](https://bugbash.antithesis.com/#about) in April—Zac as a speaker—and would love to be able to say that Hypothesis has great integration with Antithesis!)

- [x] Check whether we need to override `random.random()` too (e: we do)