@Zac-HD 2025-03-11
```
_ test_raise_invalid_argument[indexes(dtype=<class 'int'>, elements=text())] __
  | ExceptionGroup: Hypothesis found 2 distinct failures. (2 sub-exceptions)
  +-+---------------- 1 ----------------
    | Traceback (most recent call last):
    |   File ".../hypothesis/extra/pandas/impl.py", line 132, in convert_element
    |     raise InvalidArgument(
    | hypothesis.errors.InvalidArgument: Cannot convert draw(elements)='9223372036854775808' of type str to dtype <i8
    | while generating 'x' from indexes(elements=text(), dtype=int)
    +---------------- 2 ----------------
    | Traceback (most recent call last):
    |   File ".../hypothesis/extra/pandas/impl.py", line 132, in convert_element
    |     raise InvalidArgument(
    | hypothesis.errors.InvalidArgument: Cannot convert draw(elements)='' of type str to dtype <i8
    | while generating 'x' from indexes(elements=text(), dtype=int)
    +------------------------------------
```

Oh, hilarious - we're now so good at finding error-inducing strings that we'll need to set `report_multiple_bugs=False` (or handle `ExceptionGroup`s, but that's more annoying without `except*` or _very_ recent Pytest features)