@jobh 2025-03-12
I was curious about the performance impact, and indeed: the cost of 10_000 examples of `st.texts()` jumps from 4.14 to 5.90 seconds (just a single test reported, not doing proper benchmarking here).

But: when batching the `urandom` reads the backend is noticeably *faster* than the default one (4.06 →3.65 s).

I think it's worthy of note that there seems to be some performance potential in the default random generator.

For the record, the quick-and-dirty batching implementation tested was:
```
    _readahead = b''

    @classmethod
    def _urandom(cls, size: int) -> bytes:
        if len(cls._readahead) < size:
            with open("/dev/urandom", "rb") as f:
                cls._readahead += f.read(max(size, 256))
        ret = cls._readahead[:size]
        cls._readahead = cls._readahead[size:]
        return ret
```