@jobh 2025-03-12
Mind you, it doesn't really make sense that it's faster so there's probably something fishy :shrug: