@tybug 2025-03-12
I experimented several years ago (before I was ever concretely involved in Hypothesis) with batching generating randomness ahead of time and it did seem to help, so it might not be crazy. Would be interesting to experiment with a custom Random that overrides `getrandbits` to batch, say 512 bytes at a time. There are two primitive `Random` operations though, `random.random()` and `random.getrandbits`, and I wouldn't expect this to speed up the former. (But we can probably get away with getrandbits for everything except for biased booleans, even if we currently aren't..)